refactor : carMove 메소드 분리 <description> carMove 메소드를 실행하기 위해 필요한 함수 move.js로 이동
